Updating database code. Created a way to average different trials. Made UI a bit friendlier.

+++ datapoint.h
@@ -11,7 +11,7 @@ struct dataPoint
     float rxnTemp;
     float rxnHeat;
     float rxnRevCP;
-    float rxnArea;
+    int rxnArea;
 
 };
